BF: convert years to months BEFORE rounding and casting to int
Otherwise we would covert only full years, but NIH wants month-precision.

https://nda.nih.gov/data_structure.html?short_name=image03  says:
Age in months at the time of the interview/test/sampling/imaging.
Age is rounded to chronological month. If the research participant is 15-days-old at time of interview, the appropriate value would be 0 months. If the participant is 16-days-old, the value would be 1 month.